completed webhook handler

+++ checkout/webhook_handler.py
@@ -1,14 +1,20 @@
 from django.http import HttpResponse
 
+from .models import Order, OrderLineItem
+from products.models import Product
+
+import json
+import time
+
 
 class StripeWH_Handler:
-    """Handle Stripe webhooks"""
+    """ Handle Stripe webhooks """
 
     def __init__(self, request):
         self.request = request
 
     def handle_event(self, event):
-        """ Handle a generic/unknown/unexpected webhook event"""
+        """ Handle a generic/unknown/unexpected webhook event """
         return HttpResponse(
             content=f'Unhandled webhook received: {event["type"]}',
             status=200)
@@ -22,60 +28,69 @@ class StripeWH_Handler:
 
         billing_details = intent.charges.data[0].billing_details
         shipping_details = intent.shipping
-        grand_total = round(intent.data.charges[0].amount / 100, 2)
+        grand_total = round(intent.charges.data[0].amount / 100, 2)
 
-        # Clean data in the shipping details 
+        # Clean data in the shipping details
         for field, value in shipping_details.address.items():
             if value == "":
                 shipping_details.address[field] = None
 
         order_exists = False
-        try: 
-            order = Order.objects.get(
-                first_name__iexact=shipping_details.name,
-                email__iexact=shipping_details.email,
-                phone_number__iexact=shipping_details.phone,
-                country__iexact=shipping_details.country,
-                streetaddress__iexact=shipping_details.line1,
-                house_number__iexact=shipping_details.line2,
-                city__iexact=shipping_details.city,
-                grand_total= grand_total,
-            )
-            order_exists = True
-            return HttpResponse(
-                content=f'Webhook received: {event["type"]} | SUCCESS: Verified order already stored',
-                status=200)
-        except Order.DoesNotExist:
+        attempt = 1
+        while attempt <= 5:
             try:
-
+                order = Order.objects.get(
+                    first_name__iexact=shipping_details.name,
+                    email__iexact=billing_details.email,
+                    phone_number__iexact=shipping_details.phone,
+                    country__iexact=shipping_details.address.country,
+                    streetaddress__iexact=shipping_details.address.line1,
+                    house_number__iexact=shipping_details.address.line2,
+                    city__iexact=shipping_details.address.city,
+                    grand_total=grand_total,
+                    original_bag=bag,
+                    stripe_pid=pid,
+                )
+                order_exists = True
+                break
+            except Order.DoesNotExist:
+                attempt += 1
+                time.sleep(1)
+        if order_exists:
+            return HttpResponse(
+                content=f'Webhook received: {event["type"]} | SUCCESS: Verified order already in database',
+                status=200)
+        else:
+            order = None
+            try:
                 order = Order.objects.create(
                     first_name=shipping_details.name,
-                    email=shipping_details.email,
+                    email=billing_details.email,
                     phone_number=shipping_details.phone,
-                    country=shipping_details.country,
-                    streetaddress=shipping_details.line1,
-                    house_number=shipping_details.line2,
-                    city=shipping_details.city,
-                    grand_total= grand_total,
+                    country=shipping_details.address.country,
+                    streetaddress=shipping_details.address.line1,
+                    house_number=shipping_details.address.line2,
+                    city=shipping_details.address.city,
+                    original_bag=bag,
+                    stripe_pid=pid,
                 )
                 for item_id, item_data in json.loads(bag).items():
-                    product = Product.objects.get(id=item_id)
-                    if isinstance(item_data, int):
-                        order_line_item = OrderLineItem(
-                            order=order,
-                            product=product,
-                            quantity=item_data,
-                        )
-                        order_line_item.save()
+                        product = Product.objects.get(id=item_id)
+                        if isinstance(item_data, int):
+                            order_line_item = OrderLineItem(
+                                order=order,
+                                product=product,
+                                quantity=item_data,
+                            )
+                            order_line_item.save()
             except Exception as e:
                 if order:
                     order.delete()
                 return HttpResponse(
                     content=f'Webhook received: {event["type"]} | ERROR: {e}',
                     status=500)
-
         return HttpResponse(
-            content=f'Webhook received: {event["type"]}',
+            content=f'Webhook received: {event["type"]} | SUCCESS: Created order in webhook',
             status=200)
 
     def handle_payment_intent_payment_failed(self, event):
